requirements.txt
Django==3.0.8
djangorestframework==3.11.1
pytz==2020.1
asgiref==3.2.10
joblib
tensorflow
keras
opencv-python
opencv-python-headless
